refactor crystal class
